+++ src/App.tsx
@@ -1,9 +1,9 @@
-import { IssuesProvider } from './pages/IssuesPage'
 import PageRouter from './pages/PageRouter'
 
 import { ThemeProvider } from 'styled-components'
 import GlobalStyle from './styles/base/GlobalStyles'
 import { Theme } from './styles/base/DefaultTheme'
+import { IssuesProvider } from './store/IssuesContext'
 
 function App() {
   return (
